+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## 5.0.3
+
+- Added support for GitRefs, GitTag in **GitApi**.
+- Merged PR: [Can I implement deleteTag or deleteBranch method? #27](https://github.com/hkarthik7/azure-devops-java-sdk/issues/27)
 
 ## 5.0.2